Split CLI in a dedicated folder
